Module Tezos_shell_services.Injection_services

val block : Tezos_base__TzPervasives.RPC_context.#simple -> ?async:bool -> ?force:bool -> ?chain:Chain_services.chain -> Stdlib.Bytes.t -> Tezos_base.Operation.t list list -> (Tezos_crypto.Block_hash.t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

block cctxt ?async ?force raw_block tries to inject raw_block inside the node. If ?async is true, raw_block will be validated before the result is returned. If ?force is true, the block will be injected even on non strictly increasing fitness.

val operation : Tezos_base__TzPervasives.RPC_context.#simple -> ?async:bool -> ?chain:Chain_services.chain -> Stdlib.Bytes.t -> (Tezos_crypto.Operation_hash.t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val protocol : Tezos_base__TzPervasives.RPC_context.#simple -> ?async:bool -> Tezos_base.Protocol.t -> (Tezos_crypto.Protocol_hash.t, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

module S : sig ... end