Module Tezos_shell_services__Block_services

type chain = [

| `Main

| `Test

| `Hash of Tezos_crypto.Chain_id.t

]

type chain_prefix = unit * chain

val chain_path : (unit, chain_prefix) Tezos_rpc.RPC_path.t

val parse_chain : string -> (chain, string) Stdlib.result

val chain_to_string : chain -> string

val chain_arg : chain Tezos_rpc.RPC_arg.t

type block = [

| `Genesis

| `Head of int

| `Alias of [ `Caboose | `Checkpoint | `Savepoint ] * int

| `Hash of Tezos_crypto.Block_hash.t * int

| `Level of Stdlib.Int32.t

]

val parse_block : string -> (block, string) Stdlib.result

val to_string : block -> string

type prefix = (unit * chain) * block

val dir_path : (chain_prefix, chain_prefix) Tezos_rpc.RPC_path.t

val path : (chain_prefix, chain_prefix * block) Tezos_rpc.RPC_path.t

val mempool_path : ('a, 'b) Tezos_rpc.RPC_path.t -> ('a, 'b) Tezos_rpc.RPC_path.t

val live_blocks_path : ('a, 'b) Tezos_rpc.RPC_path.t -> ('a, 'b) Tezos_rpc.RPC_path.t

type operation_list_quota = {

max_size : int;

max_op : int option;

}

type raw_context =

| Key of Stdlib.Bytes.t
A leaf, containing a value

| Dir of raw_context Tezos_base__TzPervasives.TzString.Map.t
A directory, mapping keys to nested raw_contexts

| Cut
An omitted piece, because it is too deep compared to the maximum depth requested in the /chains/<chain_id>/blocks/<block_id/context/raw/bytes RPC

The low-level storage exposed as a tree

val pp_raw_context : Stdlib.Format.formatter -> raw_context -> unit

type Tezos_base__TzPervasives.error += Invalid_depth_arg of int

type merkle_hash_kind =

| Contents
The kind associated to leaves

| Node
The kind associated to directories

The kind of a merkle_node

type merkle_node =

| Hash of merkle_hash_kind * string
A shallow node: just a hash

| Data of raw_context
A full-fledged node containing actual data

| Continue of merkle_tree
An edge to a more nested tree

A node in a merkle_tree

and merkle_tree = merkle_node Tezos_base__TzPervasives.TzString.Map.t

The type of Merkle tree used by the light mode

type merkle_leaf_kind =

| Hole

| Raw_context

Whether an RPC caller requests an entirely shallow Merkle tree (Hole) or whether the returned tree should contain data at the given key (Raw_context)

val pp_merkle_node : Stdlib.Format.formatter -> merkle_node -> unit

val pp_merkle_tree : Stdlib.Format.formatter -> merkle_tree -> unit

module type PROTO = sig ... end

type protocols = {

current_protocol : Tezos_crypto.Protocol_hash.t;

next_protocol : Tezos_crypto.Protocol_hash.t;

}

val protocols : Tezos_rpc.RPC_context.#simple -> ?⁠chain:chain -> ?⁠block:block -> unit -> protocols Tezos_base__TzPervasives.tzresult Lwt.t

module Make : functor (Proto : PROTO) -> functor (Next_proto : PROTO) -> sig ... end

module Fake_protocol : PROTO

module Empty : module type of sig ... end